+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project
 adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### FIXED
+
+-   `smartstring` will now refuse to compile on 32-bit big-endian architectures, where assuming that
+    the high bit of a pointer is always empty is going to be a very bad idea.
+
 ## [0.2.3] - 2020-07-07
 
 ### ADDED